backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
greenlet==3.1.1
asyncpg==0.30.0
pgvector==0.3.6
pydantic==2.10.4
python-multipart==0.0.20
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.58.1
python-dotenv==1.0.1
alembic==1.14.0
httpx==0.28.1
PyJWT==2.10.1
websockets==14.1
